+++ gui/view_progress_bar.go
@@ -208,8 +208,8 @@ func progressBarAmendLayout(
 				Duration: 1500 * time.Millisecond,
 				Keyframes: []Keyframe{
 					{At: 0, Value: 0},
-					{At: 0.5, Value: 1, Easing: easeInOutCSS},
-					{At: 1, Value: 0, Easing: easeInOutCSS},
+					{At: 0.5, Value: 1, Easing: EaseInOutCSS},
+					{At: 1, Value: 0, Easing: EaseInOutCSS},
 				},
 				OnValue: func(v float32, w *Window) {
 					pm := StateMap[string, float32](